web/requirements.txt
asgiref==3.8.1
astroid==2.5
attrs==23.2.0
backoff==2.2.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2.24
django-address==0.2.8
django-cors-headers==4.3.1
django-extensions==3.2.3
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory-boy==2.12.0
Faker==24.4.0
geographiclib==1.52
geopy==1.21.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==3.6.0
gunicorn==20.1.0
idna==3.6
isort==4.3.21
lazy-object-proxy==1.10.0
mccabe==0.6.1
more-itertools==10.2.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==1.2.0
pathlib==1.0.1
phonenumbers==8.11.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==19.8.18
PyJWT==2.8.0
pylint==2.5.3
pytest==5.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==5.4
ratelimit==2.2.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.10.0
uwsgi==2.0.20
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.12.1
